Initial commit: - Piece : char pieceType - SquareGameboard : Piece[][] - TicTacToe : where the game is run
This initial commit is all the code I have written thus far.

A lot of the time went into figuring out how to run checks with for-loops over an Object[][] matrix and look for a winning board state and then return a winner.

Especially the bloody Columns were a pain in the ass.

All that runs now. nextUp.txt contains a ToDo list for further work.